// front/src/app/models/order-item.model.ts
// export class OrderItem {
//   OrderItemID: number | null = null;
//     OrderID: number | null = null;
//     ItemID: number | null = null;
//     Quantity: number | null = null;
//     ItemName:string ='';
//     Price:number | null = null;
//     Total:number | null = null;
// }


export class OrderItem {
  OrderItemID: number  = 0;
    OrderID: number  = 0;
    ItemID: number  = 0;
    Quantity: number  = 0;
    ItemName:string ='';
    Price:number  = 0;
    Total:number  = 0;
}